pyproject.toml
[project]
name = "magetools"
version = "1.1.0"
description = "Dynamic tool discovery for AI agents using the MCP-Zero Active Discovery pattern"
readme = "README.md"
authors = [
    { name = "Malcom Godlike", email = "ivmno1special@gmail.com" }
]
requires-python = ">=3.13"

# Core dependencies only - keeps base install light
dependencies = [
    "pyyaml>=6.0.3",
]

# Optional dependencies for full functionality
[project.optional-dependencies]
# Google GenAI provider (recommended)
google = [
    "google-genai>=1.61.0",
    "google-generativeai>=0.8.6",
]
# Vector store
vectordb = [
    "chromadb>=1.4.1",
]
# ADK integration (for agents)
adk = [
    "google-adk>=1.23.0",
    "python-dotenv>=1.2.1",
]
# Install everything: pip install magetools[full]
full = [
    "magetools[google,vectordb,adk]",
]

[project.scripts]
magetools = "magetools.__main__:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    # Include all optional deps for development
    "chromadb>=1.4.1",
    "google-adk>=1.23.0",
    "google-genai>=1.61.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "pytest-asyncio>=1.3.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
known_first_party = ["magetools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
